Fixing bug #3
The Extension now orders the images in the correct order. Please
keep in mind, that the extension default is still set to "random".
If you want to use the exact same order as in the backend, you have
to override the template.

Oh yeah, the Contao 3 models suck like hell :(

Signed-off-by: Leo Unglaub <leo.unglaub@2create.at>